+++ lib/src/cache/default_cache_delegate.dart
@@ -1,6 +1,6 @@
 import 'dart:typed_data';
 
-import 'package:ok_image/ok_image.dart';
+import 'package:ok_image/src/cache/cache_delegate.dart';
 import 'package:ok_image/src/cache/cache_manager.dart';
 
 Future<Uint8List> defaultCache(String url, DefaultFuture createDefaultFuture, {bool followRedirects}) async {
@@ -8,3 +8,12 @@ Future<Uint8List> defaultCache(String url, DefaultFuture createDefaultFuture, {b
   await icm.init();
   return icm.getImageBytes(url);
 }
+
+Uint8List getImageBytes(String url) {
+  var icm = ImageCacheManager();
+  if (!icm.isInit) return null;
+
+  return icm.getImageBytesSync(url);
+}
+
+bool isDownload(String url) => ImageCacheManager().isInit && ImageCacheManager().isDownload(url);