+++ lib/src/cache/download_error.dart
@@ -1,0 +1,7 @@
+class DownloadError extends Error {
+  String msg;
+  Error originError;
+  StackTrace stackTrace;
+
+  DownloadError(this.msg, this.originError, [this.stackTrace]);
+}