+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # CHANGELOG
 
+## [0.1.0] update request and cache
+
+Now the same URL will share a download, waiting for the download to complete and return together.
+
+### +1
+
+update readme
+
 ## [0.0.3] add cache delegate
 
 add a default cache delegate to cache image